Updates for DirectMLSuperResolution sample (#7)
* Updated to August 2019 release of DirectX Tool Kit

* Minor code review

* Markdown readme with YAML front-matter

+++ DirectMLSuperResolution/Kits/DirectXTK12/Src/CommonStates.cpp
@@ -395,7 +395,7 @@ public:
     {
         SetDebugObjectName(mDescriptors.Heap(), L"CommonStates");
 
-        for (int i = 0; i < static_cast<int>(SamplerIndex::Count); ++i)
+        for (size_t i = 0; i < static_cast<size_t>(SamplerIndex::Count); ++i)
         {
             device->CreateSampler(&SamplerDescs[i], mDescriptors.GetCpuHandle(i));
         }
